Big updae

+++ nx/common/utils.py
@@ -9,7 +9,6 @@ import unicodedata
 
 
 def unaccent(instr,encoding="utf-8"):
-    #return unicodedata.normalize('NFKD', unicode(instr,encoding)).encode('ascii', 'ignore')
     return unicodedata.normalize('NFKD', instr).encode('ascii', 'ignore')
 
 def s2time(secs):
@@ -33,10 +32,10 @@ def f2tc(f,base=25):
         f = int(f)
     except:
         return "--:--:--.--"
-    hh = (f / base) / 3600
-    mm = ((f / base) / 60) - (hh*60)
-    ss = (f/base) - (hh*3600) - (mm*60)
-    ff = f - (hh*3600*base) - (mm*60*base) - (ss*base)
+    hh = int((f / base) / 3600)
+    mm = int(((f / base) / 60) - (hh*60))
+    ss = int((f/base) - (hh*3600) - (mm*60))
+    ff = int(f - (hh*3600*base) - (mm*60*base) - (ss*base))
     return "%.2d:%.2d:%.2d:%.2d" % (hh,mm,ss,ff) 
  
  
@@ -44,12 +43,12 @@ def s2tc(s,base=25):
     """Converts seconds to timecode"""
     try:    f = int(s*base)
     except: return "--:--:--:--"
-    hh = (f / base) / 3600
-    hhd = (hh%24)
-    mm = ((f / base) / 60) - (hh*60)
-    ss = (f/base) - (hh*3600) - (mm*60)
-    ff = f - (hh*3600*base) - (mm*60*base) - (ss*base)
-    return "%.2d:%.2d:%.2d:%.2d" % (hhd,mm,ss,ff) 
+    hh  = int((f / base) / 3600)
+    hhd = int((hh%24))
+    mm  = int(((f / base) / 60) - (hh*60))
+    ss  = int((f/base) - (hh*3600) - (mm*60))
+    ff  = int(f - (hh*3600*base) - (mm*60*base) - (ss*base))
+    return "%02d:%02d:%02d:%02d" % (hhd,mm,ss,ff) 
  
  
 def s2words(s):
